+++ src/main/groovy/com/cloudogu/gitops/features/Registry.groovy
@@ -15,7 +15,7 @@ import java.nio.file.Path
 
 @Slf4j
 @Singleton
-@Order(60)
+@Order(50)
 class Registry extends Feature {
 
     /**
@@ -45,65 +45,34 @@ class Registry extends Feature {
 
     @Override
     boolean isEnabled() {
-        return true // For now, we either deploy an internal or configure an external instance
+        return config.registry['internal']
     }
 
     @Override
     void enable() {
-        
-        if (config.registry['internal']) {
-            deployRegistry()
-        }
-        
-        if (config.registry['createImagePullSecrets']) {
-            String registryUsername = config.registry['readOnlyUsername'] ?: config.registry['username']
-            String registryPassword = config.registry['readOnlyPassword'] ?: config.registry['password']
-            
-            // Name prefix is added by k8sClient
-            List exampleAppNamespaces = [ "example-apps-staging", "example-apps-production"]
-            exampleAppNamespaces.each {
-                def namespace = it as String
-                def registrySecretName = 'registry'
-                k8sClient.createImagePullSecret(registrySecretName, namespace, 
-                        config.registry['url'] as String/* Only domain matters, path would be ignored */,
-                        registryUsername, registryPassword)
-                
-                k8sClient.patch('serviceaccount', 'default', namespace,
-                        [ imagePullSecrets: [ [name: registrySecretName] ]])
 
-                if (config.registry['twoRegistries']) {
-                    k8sClient.createImagePullSecret('proxy-registry', namespace,
-                            config.registry['proxyUrl'] as String, config.registry['proxyUsername'] as String,
-                            config.registry['proxyPassword'] as String)
-                }
-            }
-        }
-
-    }
-
-    private void deployRegistry() {
         def helmConfig = config['registry']['helm']
-
+        
         Map yaml = [
                 service: [
                         nodePort: ApplicationConfigurator.DEFAULT_REGISTRY_PORT,
-                        type    : 'NodePort'
+                        type: 'NodePort'
                 ]
         ]
         log.trace("Helm yaml to be applied: ${yaml}")
         fileSystemUtils.writeYaml(yaml, tmpHelmValues.toFile())
-
+        
         if (config['registry']['internalPort'] != ApplicationConfigurator.DEFAULT_REGISTRY_PORT) {
             /* Add additional node port
                30000 is needed as a static by docker via port mapping of k3d, e.g. 32769 -> 30000 on server-0 container
                See "-p 30000" in init-cluster.sh
                e.g 32769 is needed so the kubelet can access the image inside the server-0 container
              */
-            k8sClient.createServiceNodePort('docker-registry-internal-port',
+            k8sClient.createServiceNodePort('docker-registry-internal-port', 
                     CONTAINER_PORT, config['registry']['internalPort'] as String,
                     'default')
         }
-
+        
         deployer.deployFeature(
                 helmConfig['repoURL'] as String,
                 'registry',
@@ -113,5 +82,6 @@ class Registry extends Feature {
                 'docker-registry',
                 tmpHelmValues
         )
+
     }
 }
