+++ src/main/groovy/com/cloudogu/gitops/features/ScmManager.groovy
@@ -13,7 +13,7 @@ import jakarta.inject.Singleton
 
 @Slf4j
 @Singleton
-@Order(80)
+@Order(60)
 class ScmManager extends Feature {
 
     static final String HELM_VALUES_PATH = "scm-manager/values.ftl.yaml"
